Tokens for service accounts

+++ Api/AvaluxAuth.Api/Controllers/UsersController.cs
@@ -1,5 +1,6 @@
 ﻿using AvaluxAuth.Abstractions;
 using AvaluxAuth.Api.Schemas;
+using AvaluxAuth.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
@@ -8,9 +9,9 @@ namespace AvaluxAuth.Api.Controllers;
 
 [ApiController]
 [Route("api/v1/admin/apps/{applicationId:guid}/users")]
-[Authorize(Policy = Config.AdminPolicy)]
+[Authorize(Policy = Config.AdminOrServiceAccountPolicy)]
 [EnableCors(PolicyName = Config.AdminPolicy)]
-public class UsersController(IUserRepository userRepository) : ControllerBase
+public class UsersController(IUserRepository userRepository, IUserService userService) : ControllerBase
 {
     [HttpGet]
     public async Task<ActionResult<UsersResponseSchema>> GetUsers(Guid applicationId,
@@ -18,6 +19,8 @@ public class UsersController(IUserRepository userRepository) : ControllerBase
         [FromQuery] int? limit = null,
         CancellationToken ct = default)
     {
+        if (!User.HasApplication(applicationId) || !User.HasPermission(TokenPermissions.ReadUserInfo))
+            return Unauthorized();
         var users = limit == null
             ? await userRepository.GetUsersAsync(applicationId, ct)
             : await userRepository.GetUsersAsync(applicationId, page, limit.Value, ct);
@@ -31,9 +34,39 @@ public class UsersController(IUserRepository userRepository) : ControllerBase
         });
     }
 
+    [HttpGet("{userId:guid}")]
+    public async Task<ActionResult<UserWithAccounts>> GetUser(Guid applicationId, Guid userId, CancellationToken ct)
+    {
+        if (!User.HasApplication(applicationId) || !User.HasPermission(TokenPermissions.ReadUserInfo))
+            return Unauthorized();
+        var user = await userRepository.GetUserWithAccountsAsync(userId, ct);
+        if (user == null)
+            return NotFound();
+        return Ok(user);
+    }
+
+    [HttpGet("{userId:guid}/accessToken")]
+    public async Task<ActionResult<UserWithAccounts>> GetUserAccessToken(Guid applicationId, Guid userId,
+        [FromQuery] Guid? providerId, [FromQuery] string? providerKey,
+        CancellationToken ct)
+    {
+        if (providerId == null && providerKey == null)
+            return BadRequest("providerId or providerKey is required");
+        if (!User.HasApplication(applicationId) || !User.HasPermission(TokenPermissions.ReadUserAccessToken))
+            return Unauthorized();
+        var accessToken = providerId == null
+            ? await userService.GetAccessTokenAsync(userId, providerKey!, ct)
+            : await userService.GetAccessTokenAsync(userId, providerId.Value, ct);
+        if (accessToken == null)
+            return NotFound();
+        return Ok(accessToken);
+    }
+
     [HttpDelete("{userId:guid}")]
     public async Task<ActionResult> DeleteUser(Guid applicationId, Guid userId, CancellationToken ct = default)
     {
+        if (!User.HasApplication(applicationId) || !User.HasPermission(TokenPermissions.DeleteUser))
+            return Unauthorized();
         var res = await userRepository.DeleteUserAsync(userId, ct);
         if (!res)
             return NotFound("User not found");
